+++ database/migrations/2023_02_13_134941_create_contacts_metas_table.php
@@ -13,17 +13,20 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('contacts_metas', function (Blueprint $table) {
-            $table->id();
+        if(config('filament-accounts.features.contacts')){
+            Schema::create('contacts_metas', function (Blueprint $table) {
+                $table->id();
 
-            $table->unsignedBigInteger('model_id')->nullable();
-            $table->string('model_type')->nullable();
+                $table->unsignedBigInteger('model_id')->nullable();
+                $table->string('model_type')->nullable();
 
-            $table->foreignId('contact_id')->references('id')->on('contacts')->onDelete('cascade');
-            $table->string('key')->index();
-            $table->json('value')->nullable();
-            $table->timestamps();
-        });
+                $table->foreignId('contact_id')->references('id')->on('contacts')->onDelete('cascade');
+                $table->string('key')->index();
+                $table->json('value')->nullable();
+                $table->timestamps();
+            });
+        }
+
     }
 
     /**
@@ -33,6 +36,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('contacts_metas');
+        if(config('filament-accounts.features.contacts')) {
+            Schema::dropIfExists('contacts_metas');
+        }
     }
 };
